pyproject.toml
[tool.poetry]
name = "catalystwan"
version = "0.31.0"
description = "Cisco Catalyst WAN SDK for Python"
authors = ["kagorski <kagorski@cisco.com>"]
readme = "README.md"
repository = "https://github.com/CiscoDevNet/catalystwan"

[tool.poetry.dependencies]
python = "^3.8.0"
requests = "^2.27.1"
python-dateutil = "^2.8.2"
attrs = "^21.4.0"
ciscoconfparse = "1.9.41"
tenacity = "^8.1.0"
Jinja2 = "^3.1.2"
flake8-quotes = "^3.3.1"
clint = "^0.5.1"
requests-toolbelt = "^1.0.0"
packaging = "^23.0"
pydantic = "^2.5"
typing-extensions = "^4.6.1"

[tool.poetry.dev-dependencies]
parameterized = "^0.8.1"
pytest = "^7.1.2"
pytest-mock = "^3.7.0"
pytest-subtests = "^0.11.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
mypy = "^1.0.0"
flake8 = "^5.0.4"
Sphinx = "^5.2.3"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
